fix(list): move localization to list.component.ts #5280
