+++ projects/igniteui-angular/src/lib/list/list.component.html
@@ -2,13 +2,13 @@
 
 <ng-template #defaultEmptyList>
     <article class="message">
-        {{this.resourceStrings.igx_list_no_items}}
+        {{resourceStrings.igx_list_no_items}}
     </article>
 </ng-template>
 
 <ng-template #defaultDataLoading>
     <article class="message">
-        {{this.resourceStrings.igx_list_no_items}}
+        {{resourceStrings.igx_list_no_items}}
     </article>
 </ng-template>
